+++ src/FluentSerializer.Core/Configuration/IConfigurationStack.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace FluentSerializer.Core.Configuration;
@@ -14,4 +15,13 @@ public interface IConfigurationStack<T> : IEnumerable<T>
 	/// This will replace an existing item at the current position, use <paramref name="forceTop"/> = <c>true</c> to re-add it to the top
 	/// </remarks>
 	public IConfigurationStack<T> Use(T item, bool forceTop = false);
+
+	/// <summary>
+	/// Insert this <typeparamref name="T"/> at the top of this configuration,
+	/// the <see cref="Func{TResult}"/> will be invoked immediately if called like this.
+	/// </summary>
+	/// <remarks>
+	/// This will replace an existing item at the current position, use <paramref name="forceTop"/> = <c>true</c> to re-add it to the top
+	/// </remarks>
+	public IConfigurationStack<T> Use(Func<T> item, bool forceTop = false);
 }